Add files via upload
                                           Working 
When we run our program it will first ask PIN shown in 

When you entered the PIN (5678) menu is open.
Then we have to select our require option .
26
PIN and Menu

When we select ‘B’ account balance is shown by ATM 
Balance

Similarly chose option like “D”, ”W”.
When our choice is “C” and fill what it ask, then our PIN will be change.
 describe process briefly
27
Change PIN 

In This ATM other facility is also available. For other facility we have to select option ‘O’.
How other sevices work show
When you select ‘O’, new sub menu appear .
Select Option ‘M’ for Mobile recharge.
Fill mobile number and again fill to conform. In case if it is wrong it show “This mobile number 
not exist”. In case if it is correct it ask amount.
If your balance is less then amount you top-up, then it show your current balance in not 
sufficient